refactor(api): move cache file and test
