+++ api/lib/infrastructure/caches/DistributedCache.js
@@ -1,6 +1,6 @@
+import { Cache } from '../../../src/shared/infrastructure/caches/Cache.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
 import { RedisClient } from '../../../src/shared/infrastructure/utils/RedisClient.js';
-import { Cache } from './Cache.js';
 
 class DistributedCache extends Cache {
   constructor(underlyingCache, redisUrl, channel) {
