+++ api/lib/infrastructure/caches/LayeredCache.js
@@ -1,5 +1,5 @@
+import { Cache } from '../../../src/shared/infrastructure/caches/Cache.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
-import { Cache } from './Cache.js';
 
 class LayeredCache extends Cache {
   constructor(firstLevelCache, secondLevelCache) {
