+++ api/tests/shared/unit/infrastructure/caches/Cache_test.js
@@ -1,5 +1,5 @@
-import { Cache } from '../../../../lib/infrastructure/caches/Cache.js';
-import { expect } from '../../../test-helper.js';
+import { Cache } from '../../../../../src/shared/infrastructure/caches/Cache.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Caches | Cache', function () {
   const cacheInstance = new Cache();
